+++ packages/cli/src/commands/configure/configure_profile_command.ts
@@ -1,10 +1,11 @@
 import { Argv, CommandModule } from 'yargs';
-import { AmplifyPrompter, Printer } from '@aws-amplify/cli-core';
+import { AmplifyPrompter } from '@aws-amplify/cli-core';
 import { DEFAULT_PROFILE } from '@smithy/shared-ini-file-loader';
 import { EOL } from 'os';
 import { Open } from '../open/open.js';
 import { ArgumentsKebabCase } from '../../kebab_case.js';
 import { ProfileController } from './profile_controller.js';
+import { printer } from '../../printer.js';
 
 const configureAccountUrl =
   'https://docs.amplify.aws/gen2/start/account-setup/';
@@ -44,7 +45,7 @@ export class ConfigureProfileCommand
       profileName
     );
     if (profileExists) {
-      Printer.print(
+      printer.print(
         `Profile '${profileName}' already exists!${EOL}${profileSetupInstruction}`
       );
       return;
@@ -54,7 +55,7 @@ export class ConfigureProfileCommand
     });
 
     if (!hasIAMUser) {
-      Printer.print(profileSetupInstruction);
+      printer.print(profileSetupInstruction);
 
       await Open.open(configureAccountUrl, { wait: false });
       await AmplifyPrompter.input({
@@ -73,14 +74,14 @@ export class ConfigureProfileCommand
       message: `Enter the AWS region to use with the '${profileName}' profile (eg us-east-1, us-west-2, etc):`,
     });
 
-    await this.profileController.appendAWSFiles({
+    await this.profileController.createOrAppendAWSFiles({
       profile: profileName,
       region,
       accessKeyId,
       secretAccessKey,
     });
 
-    Printer.print(`Created profile ${profileName} successfully!`);
+    printer.print(`Created profile ${profileName} successfully!`);
   };
 
   /**