+++ packages/cli/src/commands/configure/telemetry/configure_telemetry_command.test.ts
@@ -1,4 +1,3 @@
-import { Printer } from '@aws-amplify/cli-core';
 import { beforeEach, describe, it, mock } from 'node:test';
 import yargs, { CommandModule } from 'yargs';
 import assert from 'node:assert';
@@ -8,6 +7,7 @@ import {
   USAGE_DATA_TRACKING_ENABLED,
   configControllerFactory,
 } from '@aws-amplify/platform-core';
+import { printer } from '../../../printer.js';
 
 void describe('configure command', () => {
   const mockedConfigControllerSet = mock.fn();
@@ -19,18 +19,17 @@ void describe('configure command', () => {
   );
   const parser = yargs().command(telemetryCommand as unknown as CommandModule);
   const commandRunner = new TestCommandRunner(parser);
-
-  const mockedPrint = mock.method(Printer, 'print');
+  const logMock = mock.method(printer, 'log');
 
   beforeEach(() => {
-    mockedPrint.mock.resetCalls();
+    logMock.mock.resetCalls();
     mockedConfigControllerSet.mock.resetCalls();
   });
 
   void it('enable telemetry & updates local config', async () => {
     await commandRunner.runCommand(`telemetry enable`);
     assert.match(
-      mockedPrint.mock.calls[0].arguments[0],
+      logMock.mock.calls[0].arguments[0],
       /You have enabled telemetry data collection/
     );
     assert.strictEqual(
@@ -46,7 +45,7 @@ void describe('configure command', () => {
   void it('disables telemetry & updates local config', async () => {
     await commandRunner.runCommand(`telemetry disable`);
     assert.match(
-      mockedPrint.mock.calls[0].arguments[0],
+      logMock.mock.calls[0].arguments[0],
       /You have disabled telemetry data collection/
     );
     assert.strictEqual(
@@ -60,9 +59,9 @@ void describe('configure command', () => {
   });
 
   void it('if subcommand is not defined, it should list of subcommands and demandCommand', async () => {
-    await commandRunner.runCommand(`telemetry`);
+    const output = await commandRunner.runCommand(`telemetry`);
     assert.match(
-      mockedPrint.mock.calls[0].arguments[0],
+      output,
       /Not enough non-option arguments: got 0, need at least 1/
     );
     assert.strictEqual(mockedConfigControllerSet.mock.callCount(), 0);