+++ packages/cli/src/main_parser_factory.ts
@@ -16,6 +16,9 @@ export const createMainParser = (): Argv => {
   );
   const parser = yargs()
     .version(packageJson.version ?? '')
+    // This option is being used indirectly to configure the log level of the Printer instance.
+    // refer: https://github.com/aws-amplify/amplify-backend/blob/main/packages/cli/src/printer.ts
+    .options('debug', { type: 'boolean', default: false })
     .command(createGenerateCommand())
     .command(createSandboxCommand())
     .command(createPipelineDeployCommand())