+++ packages/client-config/src/client-config-contributor/graphql_client_config_contributor.test.ts
@@ -28,6 +28,7 @@ void describe('GraphqlClientConfigContributor', () => {
           userPoolId: 'stuff',
           authRegion: 'testRegion ',
           webClientId: 'clientId',
+          allowUnauthenticatedIdentities: 'true',
         },
       },
     });