+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/function.ts
@@ -1,9 +1,11 @@
 import { defineFunction, secret } from '@aws-amplify/backend';
+import { amplifySharedSecretNameKey } from '../../../shared_secret.js';
 
 export const myFunc = defineFunction({
   name: 'specialTestFunction',
   entry: './func-src/handler.ts',
   environment: {
     TEST_SECRET: secret('amazonSecret'),
+    TEST_SHARED_SECRET: secret(process.env[amplifySharedSecretNameKey] as string),
   },
 });