+++ packages/sandbox/src/sandbox_singleton_factory.ts
@@ -4,6 +4,7 @@ import { BackendDeployerFactory } from '@aws-amplify/backend-deployer';
 import { AmplifySandboxExecutor } from './sandbox_executor.js';
 import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 import { getSecretClient } from '@aws-amplify/backend-secret';
+import { Printer } from '@aws-amplify/cli-core';
 
 /**
  * Factory to create a new sandbox
@@ -13,7 +14,10 @@ export class SandboxSingletonFactory {
   /**
    * sandboxIdResolver allows sandbox to lazily load the sandbox backend id on demand
    */
-  constructor(private readonly sandboxIdResolver: BackendIdSandboxResolver) {}
+  constructor(
+    private readonly sandboxIdResolver: BackendIdSandboxResolver,
+    private readonly printer: Printer
+  ) {}
 
   /**
    * Returns a singleton instance of a Sandbox
@@ -24,9 +28,11 @@ export class SandboxSingletonFactory {
         this.sandboxIdResolver,
         new AmplifySandboxExecutor(
           BackendDeployerFactory.getInstance(),
-          getSecretClient()
+          getSecretClient(),
+          this.printer
         ),
-        new CloudFormationClient()
+        new CloudFormationClient(),
+        this.printer
       );
     }
     return this.instance;