+++ scripts/components/dependencies_validator.ts
@@ -1,6 +1,6 @@
-import fsp from 'fs/promises';
-import path from 'path';
 import { execa as _execa } from 'execa';
+import { EOL } from 'os';
+import { readPackageJson } from './package-json/package_json.js';
 
 export type DependencyRule =
   | {
@@ -23,25 +23,127 @@ type DependencyViolation = {
 };
 
 /**
- * Validates dependencies. Inspects each path, enumerates all dependencies
+ * Validates dependencies.
+ *
+ * Inspects each path, enumerates all dependencies
  * using 'npm ls --all --json', i.e. direct and transitive, and validates them
  * against the rules.
+ *
+ * Inspects dependency version declarations in package json files
+ * in order to assert consistency.
  */
 export class DependenciesValidator {
   /**
    * Creates dependency validator
+   * @param packagePaths paths of packages to validate
+   * @param dependencyRules dependency exclusion and inclusion rules
+   * @param linkedDependencies dependencies that should be versioned with the same version
+   * @param execa in order to inject execa mock in tests
    */
   constructor(
     private packagePaths: Array<string>,
     private dependencyRules: Record<string, DependencyRule>,
+    private linkedDependencies: Array<Array<string>>,
     private execa = _execa
   ) {}
 
   /**
-   * Validates whether all packages conform to dependency rules.
+   * Validates dependencies.
    * Throws if violation is found.
    */
   async validate() {
+    await this.validateDependencyDenyAndAllowListRules();
+    await this.validateDependencyVersionsConsistency();
+  }
+
+  /**
+   * Validates that all packages declare dependency using consistent version.
+   */
+  private async validateDependencyVersionsConsistency(): Promise<void> {
+    console.log('Checking dependency versions consistency');
+    const packageJsons = await Promise.all(
+      this.packagePaths.map((packagePath) => readPackageJson(packagePath))
+    );
+
+    type DependencyVersionsUsage = {
+      allVersions: Set<string>;
+      allDeclarations: Array<{
+        packageName: string;
+        version: string;
+      }>;
+    };
+
+    const dependencyVersionsUsages: Record<string, DependencyVersionsUsage> =
+      {};
+    for (const packageJson of packageJsons) {
+      [
+        packageJson.dependencies,
+        packageJson.devDependencies,
+        packageJson.peerDependencies,
+      ].forEach((dependencies) => {
+        if (dependencies) {
+          for (const dependencyName of Object.keys(dependencies)) {
+            const dependencyVersion = dependencies[dependencyName];
+            let dependencyVersionsUsage =
+              dependencyVersionsUsages[dependencyName];
+            if (!dependencyVersionsUsage) {
+              dependencyVersionsUsage = {
+                allVersions: new Set(),
+                allDeclarations: [],
+              };
+              dependencyVersionsUsages[dependencyName] =
+                dependencyVersionsUsage;
+            }
+            dependencyVersionsUsage.allVersions.add(dependencyVersion);
+            dependencyVersionsUsage.allDeclarations.push({
+              packageName: packageJson.name,
+              version: dependencyVersion,
+            });
+          }
+        }
+      });
+    }
+
+    const errors: Array<string> = [];
+    for (const dependencyName of Object.keys(dependencyVersionsUsages)) {
+      const dependencyVersionUsage = dependencyVersionsUsages[dependencyName];
+      if (dependencyVersionUsage.allVersions.size > 1) {
+        errors.push(
+          `Dependency ${dependencyName} is declared using inconsistent versions ${JSON.stringify(
+            dependencyVersionUsage.allDeclarations
+          )}`
+        );
+      }
+    }
+    for (const linkedDependencySpec of this.linkedDependencies) {
+      const allLinkedVersions: Set<string> = new Set();
+      for (const dependencyName of linkedDependencySpec) {
+        const dependencyVersionUsage = dependencyVersionsUsages[dependencyName];
+        dependencyVersionUsage.allVersions.forEach((version) =>
+          allLinkedVersions.add(version)
+        );
+      }
+
+      if (allLinkedVersions.size > 1) {
+        errors.push(
+          `Dependencies ${linkedDependencySpec.join(
+            ','
+          )} should be declared using same version, versions found ${Array.from(
+            allLinkedVersions
+          ).join(',')}`
+        );
+      }
+    }
+    if (errors.length > 0) {
+      throw new Error(errors.join(EOL));
+    }
+  }
+
+  /**
+   * Validates whether all packages conform to dependency deny and allow list rules.
+   * Throws if violation is found.
+   */
+  private async validateDependencyDenyAndAllowListRules(): Promise<void> {
     const violations: Array<DependencyViolation> = (
       await Promise.all(
         this.packagePaths.map((packagePath) =>
@@ -62,22 +164,13 @@ export class DependenciesValidator {
   }
 
   /**
-   * Reads a name from package.json located at package path.
-   */
-  private async getPackageName(packagePath: string): Promise<string> {
-    return JSON.parse(
-      (await fsp.readFile(path.join(packagePath, 'package.json'))).toString()
-    ).name;
-  }
-
-  /**
    * Checks dependencies of a package located at packagePath against
    * provided rules.
    */
   private async checkPackageDependencies(
     packagePath: string
   ): Promise<Array<DependencyViolation>> {
-    const packageName = await this.getPackageName(packagePath);
+    const packageName = (await readPackageJson(packagePath)).name;
     console.log(`Checking ${packageName} dependencies.`);
     const npmListResult = JSON.parse(
       // We're using 'npm ls' to reveal dependencies because it reveals