+++ packages/backend-data/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @aws-amplify/backend-data
 
+## 0.9.5
+
+### Patch Changes
+
+- Updated dependencies [618a2ea71]
+  - @aws-amplify/backend-output-schemas@0.5.2
+  - @aws-amplify/backend-output-storage@0.2.11
+
+## 0.9.4
+
+### Patch Changes
+
+- Updated dependencies [d087313e9]
+  - @aws-amplify/plugin-types@0.7.1
+  - @aws-amplify/backend-output-storage@0.2.10
+
+## 0.9.3
+
+### Patch Changes
+
+- 04f067837: Implement consistent dependency declaration check. Bumped dependencies where necessary.
+- Updated dependencies [04f067837]
+  - @aws-amplify/backend-output-schemas@0.5.1
+  - @aws-amplify/backend-output-storage@0.2.9
+
+## 0.9.2
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.2.8
+
 ## 0.9.1
 
 ### Patch Changes